Add empty token array on subject creation
This fixes a bug where password reset was failing between app start up
and first user log in.

The db-info contains a key :subject-id-to-tokens-data-id that we
expect is mandatory but was not set until the first user login. This
resulted in errors if something (such as storing a new password after
password reset) was to happen after app startup but before anyone had
logged in.